Add files via upload
Feel free to explore the project's development by watching videos detailing the construction and functionality of the application. Keep in mind that the project is a work in progress, and my last focus was on determining the percentage score for each half of the qubits. I specifically aimed to identify whether the first or second half had more matches. I'm still uncertain if I've successfully resolved that aspect, so your insights and feedback would be valuable.
https://www.youtube.com/watch?v=ZdAj_W4AsmM&t=4s p1
https://www.youtube.com/watch?v=4k3xkHoGu98&t=218s p2 also 2 more parts on youtube